+++ lib/middleware/index.ts
@@ -1,3 +1,4 @@
+export *                from './Auth';
 export {CacheControl}   from './CacheControl';
 export {Cors}           from './Cors';
 export {Security}       from './Security';